% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_possibilitylab.R
\docType{data}
\name{colors_possibilitylab}
\alias{colors_possibilitylab}
\title{Possibility Lab named color list}
\format{
An object of class \code{character} of length 7.
}
\usage{
colors_possibilitylab
}
\description{
Vectors of hex codes for colors to be used in visualizations
}
\keyword{datasets}
