% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_possibilitylab.R
\docType{data}
\name{palette_possibilitylab}
\alias{palette_possibilitylab}
\title{Possibility Lab color palette list}
\format{
An object of class \code{list} of length 2.
}
\usage{
palette_possibilitylab
}
\description{
Various color palettes for applying Possibility Lab colors to visualizations
}
\keyword{datasets}
