% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possibility_add_grid.R
\name{possibility_add_grid}
\alias{possibility_add_grid}
\title{Add missing major gridlines from default theme}
\usage{
possibility_add_grid(add_major = "x", add_minor = NULL)
}
\arguments{
\item{add_major}{adds major gridlines to specified axes. One of "x", "y", "both" or NULL}

\item{add_minor}{adds minor gridlines to specified axes. One of "x", "y", "both" or NULL}
}
\value{
a theme layer with major and/or minor axes
}
\description{
Add missing major gridlines from default theme
}
