% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possibility_defaults.R
\name{possibility_defaults}
\alias{possibility_defaults}
\title{Set Possibility Lab theme as defaults}
\usage{
possibility_defaults(
  type = "default",
  base_size = 12,
  base_family = "sourcesanspro"
)
}
\arguments{
\item{type}{one of "default", "map", "likert"}

\item{base_size}{base font size}

\item{base_family}{base font family}
}
\value{
sets the default ggplot2 theme to the Possibility Lab theme
}
\description{
Set Possibility Lab theme as defaults
}
