% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_possibilitylab.R
\name{possibility_palette}
\alias{possibility_palette}
\title{Get a color palette from Possibility Lab's palettes}
\usage{
possibility_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{the name of the palette}

\item{n}{number of colors required, up to a max}

\item{type}{discrete or continuous}
}
\value{
an object of class "palette"
}
\description{
Get a color palette from Possibility Lab's palettes
}
