% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_possibilitylab.R
\docType{data}
\name{possibility_palettes}
\alias{possibility_palettes}
\title{Possibility Lab named palette list}
\format{
An object of class \code{list} of length 7.
}
\usage{
possibility_palettes
}
\description{
List of palettes to be used in visualizations
}
\keyword{datasets}
