% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possibility_save.R
\name{possibility_save}
\alias{possibility_save}
\title{Save a plot with Possibility Lab theme and export settings}
\usage{
possibility_save(
  file,
  x = ggplot2::last_plot(),
  preset_size = c("small", "medium", "large"),
  orientation = c("horizontal", "vertical"),
  dpi = 300
)
}
\arguments{
\item{file}{Destination filepath}

\item{x}{plot object to save. Defaults to \code{last_plot()}}

\item{preset_size}{One of "small", "medium", or "large".}

\item{orientation}{One of "horizontal" (the default) or "vertical"}
}
\value{
a saved plot
}
\description{
Save a plot with Possibility Lab theme and export settings
}
