% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_possibilitylab.R
\name{scale_color_discrete}
\alias{scale_color_discrete}
\title{Discrete color scale with Possibility Lab color palette}
\usage{
scale_color_discrete(...)
}
\arguments{
\item{...}{other arguments passed to scale_color_discrete}
}
\description{
Discrete color scale with Possibility Lab color palette
}
