% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_possibilitylab.R
\name{scale_color_gradientn}
\alias{scale_color_gradientn}
\title{Gradient color scale with Possibility Lab color palette}
\usage{
scale_color_gradientn(
  ...,
  colors = c("#4097c9", "#3181af", "#286a8f", "#265b7a", "#1f526f", "#163b50", "#0d2330",
    "#040c10"),
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = "colourbar"
)
}
\arguments{
\item{...}{other arguments passed to scale_color_gradientn}

\item{colors}{colors to use}

\item{values}{NuLL}

\item{space}{must be "Lab"}

\item{na.value}{sets NA values to grey by default}

\item{guide}{uses the "colourbar" guide for gradient legends}
}
\description{
Gradient color scale with Possibility Lab color palette
}
