% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_possibilitylab.R
\name{scale_color_ordinal}
\alias{scale_color_ordinal}
\title{Ordinal color scale with Possibility Lab color palette}
\usage{
scale_color_ordinal(...)
}
\arguments{
\item{...}{other arguments passed to scale_color_ordinal}
}
\description{
Ordinal color scale with Possibility Lab color palette
}
