% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_possibilitylab.R
\name{scale_colour_discrete}
\alias{scale_colour_discrete}
\title{Discrete colour scale with Possibility Lab colour palette}
\usage{
scale_colour_discrete(...)
}
\arguments{
\item{...}{other arguments passed to scale_colour_discrete}
}
\description{
Discrete colour scale with Possibility Lab colour palette
}
