% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_possibilitylab.R
\name{scale_colour_ordinal}
\alias{scale_colour_ordinal}
\title{Ordinal colour scale with Possibility Lab colour palette}
\usage{
scale_colour_ordinal(...)
}
\arguments{
\item{...}{other arguments passed to scale_colour_ordinal}
}
\description{
Ordinal colour scale with Possibility Lab colour palette
}
