% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_possibilitylab.R
\name{scale_fill_discrete}
\alias{scale_fill_discrete}
\title{Discrete fill scale with Possibility Lab color palette}
\usage{
scale_fill_discrete(...)
}
\arguments{
\item{...}{other arguments passed to scale_fill_discrete}
}
\description{
Discrete fill scale with Possibility Lab color palette
}
