% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_possibilitylab.R
\name{scale_fill_ordinal}
\alias{scale_fill_ordinal}
\title{Ordinal fill scale with Possibility Lab color palette}
\usage{
scale_fill_ordinal(...)
}
\arguments{
\item{...}{other arguments passed to scale_fill_ordinal}
}
\description{
Ordinal fill scale with Possibility Lab color palette
}
