% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_possibilitylab.R
\name{theme_possibilitylab}
\alias{theme_possibilitylab}
\title{Apply a general Possibility Lab theme to ggplot visualizations}
\usage{
theme_possibilitylab(
  base_size = 8.5,
  base_family = "Source Sans Pro",
  base_line_size = 0.5,
  base_rect_size = 0.5
)
}
\arguments{
\item{base_size, base_family}{base font size and family}

\item{base_line_size}{base line width}

\item{base_rect_size}{base rect width}
}
\value{
a theme layer
}
\description{
Apply a general Possibility Lab theme to ggplot visualizations
}
