% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_possibilitylab.R
\name{use_possibililtylab_pal}
\alias{use_possibililtylab_pal}
\title{Use Possibility Lab color palettes}
\usage{
use_possibililtylab_pal(palette = "categorical")
}
\arguments{
\item{palette}{one of "categorical" or "sequential"}
}
\value{
Possibility Lab color palette
}
\description{
Use Possibility Lab color palettes
}
